fix: unblock main — resolve ban-scanner violations from concurrent merges
Three issues from concurrent owner merges broke the build on main:

1. 10k-line gate (src/terms/sae/manifold/tests.rs at 10119 lines):
   Extracted the OLMo real-data test group (read_npy_f32_2d,
   real_data_torus_seed_term, olmo_real_curvature_anchor_is_positive_definite,
   olmo_real_outer_fit_does_not_pin_at_collapse_sentinel) into a new
   real_data_tests.rs sibling module. tests.rs is now 9650 lines.

2. debug_assert! ban (src/families/custom_family/assembly.rs:1115):
   The #1395 logdet-collapse guard used debug_assert!, banned by the
   ban-scanner. Removed it — the log::error! already surfaces the regression
   (an unconditional assert! would panic production on a numerical guard).
   Updated the comment to explain the policy-compliant approach.

3. dead code from #1440 (theta_hvp_matrix_free in per_atom_efs.rs):
   After #1440 removed the FD fallback, the function had no production
   callers (only same-file tests). Moved it into the test module as
   pub(crate). Also removed the now-unused probe_step_for re-export from
   manifold/mod.rs.

Build: cargo +1.93.0 check --lib --tests passes (ban-scanner clean).
Tests: per_atom_efs 6/6 pass.
