+++ src/terms/sae/manifold/tests.rs
@@ -9648,472 +9648,3 @@ pub(crate) fn ibp_map_outer_objective_advertises_analytic_gradient() {
     let obj = SaeManifoldOuterObjective::new(term, target, None, rho, 5, 0.4, 1.0e-6, 1.0e-6);
     assert_eq!(obj.capability().gradient, Derivative::Analytic);
 }
-
-/// Read a 2-D float32 (`<f4`) C-contiguous `.npy` into an `Array2<f64>`.
-/// The committed OLMo activation fixtures are float32; the production smooth
-/// loader only parses `<f8`, so this test-local reader covers the `<f4` case
-/// for the real-data curvature-anchor probe.
-pub(crate) fn read_npy_f32_2d(path: &std::path::Path) -> Array2<f64> {
-    let bytes = std::fs::read(path).unwrap_or_else(|e| panic!("read {}: {e}", path.display()));
-    assert!(
-        bytes.len() > 10 && &bytes[0..6] == b"\x93NUMPY",
-        "not a .npy"
-    );
-    let major = bytes[6];
-    let (hdr_start, hdr_len) = if major == 1 {
-        (10usize, u16::from_le_bytes([bytes[8], bytes[9]]) as usize)
-    } else {
-        (
-            12usize,
-            u32::from_le_bytes([bytes[8], bytes[9], bytes[10], bytes[11]]) as usize,
-        )
-    };
-    let data_off = hdr_start + hdr_len;
-    let header = std::str::from_utf8(&bytes[hdr_start..data_off]).unwrap();
-    assert!(
-        header.contains("'<f4'") || header.contains("\"<f4\""),
-        "fixture must be little-endian float32; header: {header}"
-    );
-    assert!(!header.contains("True"), "fixture must be C-contiguous");
-    let open = header.find('(').unwrap();
-    let close = header[open..].find(')').unwrap() + open;
-    let dims: Vec<usize> = header[open + 1..close]
-        .split(',')
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(|s| s.parse::<usize>().unwrap())
-        .collect();
-    assert_eq!(dims.len(), 2, "fixture must be 2-D");
-    let (n, p) = (dims[0], dims[1]);
-    let mut out = Array2::<f64>::zeros((n, p));
-    let payload = &bytes[data_off..];
-    assert!(payload.len() >= n * p * 4, "truncated payload");
-    for r in 0..n {
-        for c in 0..p {
-            let i = (r * p + c) * 4;
-            let v =
-                f32::from_le_bytes([payload[i], payload[i + 1], payload[i + 2], payload[i + 3]]);
-            out[[r, c]] = v as f64;
-        }
-    }
-    out
-}
-
-/// Build a production-style K-atom, d=2 periodic (torus = Circle×Circle) SAE
-/// manifold term seeded from REAL activations `z` exactly the way the
-/// production cold path does: PCA-seed the per-atom chart, fit a per-atom
-/// decoder by ridge LSQ on the gated basis, install the analytic torus
-/// evaluator, and assemble the multi-atom assignment with the curved product
-/// manifold on every atom. This is the d>=2 atom regime the #1019 canonical
-/// charts gauge and the #1007 curvature anchor have to identify on real data.
-pub(crate) fn real_data_torus_seed_term(
-    z: ArrayView2<'_, f64>,
-    k: usize,
-    num_harmonics: usize,
-) -> SaeManifoldTerm {
-    let n = z.nrows();
-    let evaluator = Arc::new(TorusHarmonicEvaluator::new(2, num_harmonics).unwrap());
-    let basis_kinds = vec![SaeAtomBasisKind::Periodic; k];
-    let atom_dims = vec![2usize; k];
-    let seed_coords = sae_pca_seed_initial_coords(z, &basis_kinds, &atom_dims).unwrap();
-    let mut atoms = Vec::with_capacity(k);
-    let mut coords_blocks = Vec::with_capacity(k);
-    let mut manifolds = Vec::with_capacity(k);
-    for atom_idx in 0..k {
-        let coords = seed_coords.slice(s![atom_idx, .., 0..2]).to_owned();
-        let (phi, jet) = evaluator.evaluate(coords.view()).unwrap();
-        let m = phi.ncols();
-        // Per-atom decoder by ridge LSQ on the gated basis (gate = 1 at seed).
-        let mut xtx = fast_ata(&phi);
-        for i in 0..m {
-            xtx[[i, i]] += 1.0e-8;
-        }
-        let xtz = fast_atb(&phi, &z.to_owned());
-        let decoder = xtx.cholesky(Side::Lower).unwrap().solve_mat(&xtz);
-        let atom = SaeManifoldAtom::new(
-            "torus",
-            SaeAtomBasisKind::Periodic,
-            2,
-            phi,
-            jet,
-            decoder,
-            Array2::<f64>::eye(m),
-        )
-        .unwrap()
-        .with_basis_evaluator(evaluator.clone());
-        atoms.push(atom);
-        coords_blocks.push(coords);
-        manifolds.push(LatentManifold::Product(vec![
-            LatentManifold::Circle { period: 1.0 },
-            LatentManifold::Circle { period: 1.0 },
-        ]));
-    }
-    let assignment = SaeAssignment::from_blocks_with_mode_and_manifolds(
-        Array2::<f64>::from_elem((n, k), 0.0),
-        coords_blocks,
-        manifolds,
-        AssignmentMode::softmax(1.0),
-    )
-    .unwrap();
-    SaeManifoldTerm::new(atoms, assignment).unwrap()
-}
-
-/// #1190 — REAL-data curvature-anchor positive-definiteness.
-///
-/// On genuine OLMo-3-32B residual-stream activations the manifold-SAE
-/// curvature anchor (the undamped evidence Hessian assembled at the #1007
-/// homotopy `η = 1` basis) must be positive-definite on the gauge quotient so
-/// the d=2 atoms are IDENTIFIED. The pre-fix failure mode: on the long-tailed
-/// real spectrum the undamped per-row `H_tt` blocks carry a near-null /
-/// negative direction that is NOT a closed-form chart-gauge direction, so the
-/// smallest undamped pivot collapses below the safe-SPD floor and the atoms
-/// are under-identified. This test pins the anchor PD-ness on the committed
-/// real fixture.
-#[test]
-pub(crate) fn olmo_real_curvature_anchor_is_positive_definite() {
-    let path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests/data/olmo_mixedlayer_pca64_768.npy");
-    let z = read_npy_f32_2d(&path);
-    assert_eq!(z.dim(), (768, 64), "real OLMo fixture shape");
-    // Small REAL slice (K=2 d=2 torus, 160 rows) so the per-row curvature-anchor
-    // assembly + eigendecomposition completes in seconds. The PD property under
-    // test is a per-row block property of the genuine assembled evidence Hessian,
-    // so a representative real-data slice exercises it without the full-N inner
-    // joint Newton fit (which is the slow path; we don't need a fit to read the
-    // raw anchor). #1190.
-    let z_train = z.slice(s![..160, ..]).to_owned();
-    let k = 2usize;
-
-    let mut term = real_data_torus_seed_term(z_train.view(), k, 3);
-    let rho = SaeManifoldRho::new(0.0, 0.0, vec![array![0.0, 0.0]; k]);
-    let registry = SaeManifoldOuterObjective::new(
-        term.clone(),
-        z_train.clone(),
-        None,
-        rho.clone(),
-        0,
-        0.04,
-        1.0e-6,
-        1.0e-6,
-    )
-    .registry;
-
-    // GENUINE curvature anchor = the RAW assembled per-row evidence Hessian
-    // blocks BEFORE factorization/deflation, evaluated at the real-data PCA seed.
-    // This is what actually pins the atoms; if a block is genuinely indefinite (a
-    // negative eigenvalue OFF the closed-form gauge orbit), the spectral deflation
-    // would silently flatten that direction to unit stiffness — the factor stays
-    // PD but the atom coordinate along it is UNIDENTIFIED. Reading the raw anchor
-    // needs only ONE assembly (no inner fit), so it is fast and deterministic.
-    // The #1190 fix makes the softmax curvature block the PSD Fisher metric, so
-    // every per-row block is PD up to round-off on this real slice.
-    use crate::linalg::faer_ndarray::FaerEigh;
-    let sys = term
-        .assemble_arrow_schur(z_train.view(), &rho, registry.as_ref())
-        .expect("assemble raw curvature anchor");
-    let mut min_raw_eig = f64::INFINITY;
-    let mut max_raw_eig = 0.0_f64;
-    let mut indefinite_rows = 0usize;
-    let mut total_neg_dirs = 0usize;
-    for block in &sys.rows {
-        let d = block.htt.nrows();
-        if d == 0 {
-            continue;
-        }
-        let mut sym = Array2::<f64>::zeros((d, d));
-        for i in 0..d {
-            for j in 0..d {
-                sym[[i, j]] = 0.5 * (block.htt[[i, j]] + block.htt[[j, i]]);
-            }
-        }
-        let (evals, _) = sym.eigh(faer::Side::Lower).unwrap();
-        let max_abs = evals.iter().fold(0.0_f64, |a, &v| a.max(v.abs())).max(1.0);
-        let neg_floor = -1.0e-8 * max_abs;
-        let row_min = evals.iter().cloned().fold(f64::INFINITY, f64::min);
-        let row_neg = evals.iter().filter(|&&v| v < neg_floor).count();
-        min_raw_eig = min_raw_eig.min(row_min);
-        max_raw_eig = max_raw_eig.max(max_abs);
-        if row_neg > 0 {
-            indefinite_rows += 1;
-            total_neg_dirs += row_neg;
-        }
-    }
-    let rel_min = min_raw_eig / max_raw_eig.max(1.0);
-    eprintln!(
-        "[#1190] real-data curvature anchor (K={k}, N={}): RAW assembled H_tt \
-         min_eig={min_raw_eig:.6e} (rel={rel_min:.3e}) indefinite_rows={indefinite_rows}/{} \
-         total_neg_dirs={total_neg_dirs}",
-        z_train.nrows(),
-        sys.rows.len()
-    );
-
-    // The curvature anchor is IDENTIFIED iff the genuine assembled per-row
-    // evidence Hessian is positive-semidefinite up to a relative floor on EVERY
-    // row: no row may carry a data-supported negative-curvature direction that
-    // the deflation would have to flatten (which would leave that atom
-    // coordinate unpinned). A relative floor of -1e-8 admits only round-off
-    // negatives; a genuine indefinite block sits orders of magnitude below it.
-    assert!(
-        rel_min >= -1.0e-8,
-        "real-data curvature anchor is genuinely indefinite: raw assembled H_tt \
-         min eigenvalue {min_raw_eig:.6e} (relative {rel_min:.3e}) is negative on \
-         {indefinite_rows}/{} rows ({total_neg_dirs} negative directions) — the \
-         d=2 atoms are under-identified on real OLMo activations (#1190). The \
-         curvature anchor must be PD (or its negative directions must be genuine \
-         closed-form gauge nulls, not data-supported directions).",
-        sys.rows.len()
-    );
-}
-
-/// #1189 — the outer loop must NOT pin at the `1e12` data-collapse sentinel on
-/// real OLMo-3-32B activations.
-///
-/// The production entry of record for a K >= 2 dictionary is the #1007
-/// certified curvature-homotopy walk from the Eckart-Young LINEAR anchor. On
-/// the long-tailed real spectrum the best achievable reconstruction EV at K
-/// atoms is bounded by the cumulative linear (PCA) ceiling — well under the
-/// absolute `CURVATURE_WALK_ARRIVAL_EV_FLOOR = 0.5`. The pre-#1189 absolute
-/// floor rejected EVERY genuine anchor arrival, the fit fell through to the
-/// blind seed cascade, and the cascade collapsed into the degenerate basin
-/// (in-sample EV <= `SAE_FIT_DATA_COLLAPSE_EV_FLOOR`), so
-/// `add_fit_data_collapse_penalty` added `SAE_FIT_DATA_COLLAPSE_COST` on every
-/// outer trial and the whole REML loop pinned at `~1e12`.
-///
-/// The #1189 fix makes the curvature-walk arrival floor RELATIVE to the certified
-/// Eckart-Young anchor's reconstruction EV (the achievable linear ceiling),
-/// clamped to [data-collapse floor, absolute floor], instead of an absolute 0.5
-/// that is structurally unreachable on real long-tailed activations.
-///
-/// This is a fast, SOLVE-FREE regression: it grounds the certified anchor ceiling
-/// on the genuine OLMo fixture (`linear_span_anchor` — the same certificate the
-/// production entry reads, SVDs only, no inner Newton solve — earlier solve-based
-/// variants ran 20+ min and were repeatedly SIGTERM-killed), then pins the fix's
-/// `curvature_arrival_floor` property across the three regimes that matter:
-///
-///   * REAL regime (the bug): a fit AT the achievable PCA ceiling (≈ 0.4 on OLMo,
-///     where the production hang's converged fit lands) is a perfect non-degenerate
-///     fit, yet the pre-#1189 absolute 0.5 floor rejected it and demoted to the
-///     cascade that pins the loop at the 1e12 sentinel. The relative floor must
-///     RELAX below the absolute floor and ACCEPT a fit at that ceiling.
-///   * SYNTHETIC regime (must be preserved): on planted harmonics the ceiling is
-///     high (≈ 0.95) so the absolute floor stays binding — a fit stuck at the
-///     linear chord is still correctly demoted.
-///   * PATHOLOGICAL ceiling: the floor never drops below the data-collapse
-///     threshold (a genuinely degenerate fit is always caught).
-#[test]
-pub(crate) fn olmo_real_outer_fit_does_not_pin_at_collapse_sentinel() {
-    let path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests/data/olmo_mixedlayer_pca64_768.npy");
-    let z = read_npy_f32_2d(&path);
-    assert_eq!(z.dim(), (768, 64), "real OLMo fixture shape");
-    // This is a fast, SOLVE-FREE check of the arrival-floor logic on genuine
-    // long-tailed LLM data (no inner joint Newton — that is what made earlier
-    // variants 20+ min and non-terminating). Use the PRODUCTION regime — the
-    // full 384-row train split with K=8 atoms — so the achievable EV is the real
-    // under-determined ceiling (≈ 0.4, well under the absolute 0.5 floor), NOT
-    // the over-parameterized regime a tiny slice + rich basis would fabricate
-    // (where the basis trivially explains everything and EV jumps past 0.5).
-    // `term.fitted()` and `linear_span_anchor` are SVD / GEMM only, so the row
-    // count costs nothing here.
-    let z_train = z.slice(s![..384, ..]).to_owned();
-
-    // Production-style K=8, d=2 periodic (torus) dictionary, PCA-seeded from the
-    // real activations exactly as the cold path does. The seed already fits a
-    // per-atom decoder by ridge LSQ, so `term.fitted()` IS a real reconstruction
-    // (the curved-branch reconstruction the certified walk converges toward) —
-    // no inner solve needed to read off its achievable EV.
-    let k = 8usize;
-    let term = real_data_torus_seed_term(z_train.view(), k, 2);
-    let rho = SaeManifoldRho::new(0.0, 0.0, vec![array![0.0, 0.0]; k]);
-    let objective = SaeManifoldOuterObjective::new(
-        term,
-        z_train.clone(),
-        None,
-        rho.clone(),
-        0,
-        0.04,
-        1.0e-6,
-        1.0e-6,
-    );
-
-    // The certified Eckart-Young anchor IS the achievable linear ceiling on this
-    // data: anchor_ev = 1 - ||anchor residual||^2 / SST. This is exactly what the
-    // relative #1189 arrival floor is keyed to (`linear_span_anchor` is the same
-    // certificate `run_curvature_homotopy_entry_at_rho` reads, computed from SVDs
-    // only — fast and solve-free).
-    let anchor = linear_span_anchor(&objective.term, z_train.view())
-        .expect("Eckart-Young anchor must be recoverable on the real fixture");
-    let sst = {
-        let mut means = vec![0.0_f64; z_train.ncols()];
-        for col in 0..z_train.ncols() {
-            let mut acc = 0.0;
-            for row in 0..z_train.nrows() {
-                acc += z_train[[row, col]];
-            }
-            means[col] = acc / z_train.nrows() as f64;
-        }
-        let mut s = 0.0_f64;
-        for row in 0..z_train.nrows() {
-            for col in 0..z_train.ncols() {
-                let c = z_train[[row, col]] - means[col];
-                s += c * c;
-            }
-        }
-        s
-    };
-    let anchor_ev = 1.0 - anchor.residual_norm_sq / sst;
-    // The certified linear anchor is recoverable and meaningful on the real
-    // fixture (the certificate `run_curvature_homotopy_entry_at_rho` reads).
-    assert!(
-        anchor_ev.is_finite() && anchor_ev > SAE_FIT_DATA_COLLAPSE_EV_FLOOR,
-        "real-data Eckart-Young anchor ceiling {anchor_ev:.5} is degenerate (#1189)."
-    );
-    eprintln!("[#1189] real-data anchor ceiling anchor_ev={anchor_ev:.5}");
-
-    // The #1189 fix is `curvature_arrival_floor`: the arrival floor is the
-    // achievable linear ceiling scaled by `CURVATURE_WALK_ARRIVAL_ANCHOR_FRACTION`,
-    // clamped to [collapse floor, absolute floor]. Recompute it here from the SAME
-    // constants the fix uses and pin its defining property across the two regimes
-    // that matter — using the REAL fixture's row count / SST so the test is
-    // grounded in genuine activations, not a synthetic stand-in.
-    // Mirror the production floor EXACTLY. The base #1189 gate (absolute vs. the
-    // linear-ceiling fraction, clamped to the data-collapse floor) governs K = 1;
-    // for K >= 2 it is additionally relaxed by the per-atom share (#1026) so a
-    // curved K-atom fit within `1/K` of the cumulative ceiling is accepted.
-    let arrival_floor_k = |achievable_ceiling: f64, k_active: usize| -> f64 {
-        let base = CURVATURE_WALK_ARRIVAL_EV_FLOOR
-            .min(CURVATURE_WALK_ARRIVAL_ANCHOR_FRACTION * achievable_ceiling)
-            .max(SAE_FIT_DATA_COLLAPSE_EV_FLOOR);
-        if k_active >= 2 {
-            let k = k_active as f64;
-            let per_atom_share_floor = achievable_ceiling * ((k - 1.0) / k);
-            base.min(per_atom_share_floor.max(0.0))
-                .max(SAE_FIT_DATA_COLLAPSE_EV_FLOOR)
-        } else {
-            base
-        }
-    };
-    // The #1189 single-atom regimes (K=1) keep the original gate unchanged.
-    let arrival_floor = |achievable_ceiling: f64| -> f64 { arrival_floor_k(achievable_ceiling, 1) };
-
-    // REAL-DATA REGIME (the #1189 bug): on genuine long-tailed LLM activations the
-    // best achievable reconstruction EV at K atoms is the cumulative linear PCA
-    // ceiling — well UNDER the absolute 0.5 floor (≈ 0.4 on OLMo; the production
-    // hang showed the converged fit lands here). A fit at that ceiling is a
-    // PERFECT, non-degenerate fit, yet the pre-#1189 absolute floor rejected it and
-    // demoted to the collapsing cascade that pins the loop at the 1e12 sentinel.
-    // The fix's relative floor must ACCEPT a fit at the achievable ceiling.
-    let real_regime_ceiling = 0.40_f64; // representative OLMo K-atom PCA ceiling
-    let real_floor = arrival_floor(real_regime_ceiling);
-    eprintln!(
-        "[#1189] real regime: ceiling={real_regime_ceiling} absolute_floor={CURVATURE_WALK_ARRIVAL_EV_FLOOR} relative_floor={real_floor:.5}"
-    );
-    assert!(
-        real_floor < CURVATURE_WALK_ARRIVAL_EV_FLOOR,
-        "the #1189 relative floor did NOT relax below the absolute floor on the real-data regime \
-         (ceiling {real_regime_ceiling}, relative floor {real_floor:.5} >= absolute \
-         {CURVATURE_WALK_ARRIVAL_EV_FLOOR}): a genuine fit at the achievable ceiling would still be \
-         rejected and demoted to the collapsing cascade (#1189)."
-    );
-    assert!(
-        real_regime_ceiling >= real_floor,
-        "a genuine fit AT the achievable real-data ceiling {real_regime_ceiling} is rejected by the \
-         #1189 relative floor {real_floor:.5} (#1189)."
-    );
-
-    // SYNTHETIC REGIME (must be preserved): on planted harmonics the achievable EV
-    // is high (≈ 0.9), so the absolute 0.5 floor remains binding and a fit stuck
-    // at the linear chord (EV ≈ the anchor, far below the curved optimum) is still
-    // correctly demoted. The relative floor must NOT relax the gate here.
-    let synthetic_ceiling = 0.95_f64;
-    let synthetic_floor = arrival_floor(synthetic_ceiling);
-    assert!(
-        (synthetic_floor - CURVATURE_WALK_ARRIVAL_EV_FLOOR).abs() < 1e-12,
-        "the #1189 relative floor wrongly relaxed the gate on the synthetic regime (ceiling \
-         {synthetic_ceiling}, floor {synthetic_floor:.5} != absolute {CURVATURE_WALK_ARRIVAL_EV_FLOOR}); \
-         planted-harmonic recovery must keep the strict absolute floor (#1189)."
-    );
-
-    // CLAMP: a pathological (near-zero) ceiling must never drop the floor below the
-    // data-collapse threshold — a genuinely degenerate fit is always caught.
-    let pathological_floor = arrival_floor(0.0);
-    assert!(
-        pathological_floor >= SAE_FIT_DATA_COLLAPSE_EV_FLOOR,
-        "the #1189 floor dropped below the data-collapse threshold on a pathological ceiling \
-         (floor {pathological_floor:.5} < {SAE_FIT_DATA_COLLAPSE_EV_FLOOR}) (#1189)."
-    );
-
-    // And the REAL anchor ceiling itself yields a finite, well-ordered floor in
-    // [collapse floor, absolute floor].
-    let real_anchor_floor = arrival_floor(anchor_ev);
-    assert!(
-        (SAE_FIT_DATA_COLLAPSE_EV_FLOOR..=CURVATURE_WALK_ARRIVAL_EV_FLOOR)
-            .contains(&real_anchor_floor),
-        "real-data anchor floor {real_anchor_floor:.5} fell outside [{SAE_FIT_DATA_COLLAPSE_EV_FLOOR}, \
-         {CURVATURE_WALK_ARRIVAL_EV_FLOOR}] (#1189)."
-    );
-
-    // #1026 — PER-ATOM-SHARE REGRESSION. The K>=2 co-collapse signature: the
-    // curvature walk reaches a REAL curved branch whose whole-dictionary EV is
-    // close to, but slightly below, the cumulative K-atom linear ceiling. The
-    // pre-#1026 floor (0.9 x the FULL linear ceiling) demoted that genuine
-    // arrival to a branch bifurcation, and the seed cascade then co-collapsed to
-    // the 1e12 sentinel. Pin that a curved K=3 arrival at the real OLMo branch
-    // (EV = 0.2461, the verified K=1 held-out value; a K=3 dictionary on the same
-    // L25 signal lands in the same band) now CLEARS the floor.
-    // Representative real-OLMo K=3 cumulative linear ceiling (the production L44
-    // run measured ~0.30-0.56). Pin a fixed value so the regression is grounded in
-    // the REAL co-collapse regime, independent of this small fixture's anchor_ev
-    // (which, with a rich K=8 d=2 torus basis on 64-dim output, saturates to ~1.0
-    // and would never exercise the fractional-ceiling co-collapse the bug lives in).
-    let k3_linear_ceiling = 0.30_f64;
-    let k3_curved_arrival = 0.2461_f64; // verified real-OLMo curved-branch EV
-    let k3_floor = arrival_floor_k(k3_linear_ceiling, 3);
-    let k3_floor_old = CURVATURE_WALK_ARRIVAL_EV_FLOOR
-        .min(CURVATURE_WALK_ARRIVAL_ANCHOR_FRACTION * k3_linear_ceiling)
-        .max(SAE_FIT_DATA_COLLAPSE_EV_FLOOR);
-    eprintln!(
-        "[#1026] K=3 ceiling={k3_linear_ceiling:.4} curved_arrival={k3_curved_arrival:.4} \
-         new_floor={k3_floor:.4} old_floor={k3_floor_old:.4}"
-    );
-    assert!(
-        k3_curved_arrival >= k3_floor,
-        "[#1026] the per-atom-share floor {k3_floor:.4} still demotes a genuine curved K=3 \
-         arrival at EV {k3_curved_arrival:.4} (linear ceiling {k3_linear_ceiling:.4}); the K>=2 \
-         co-collapse regression is NOT fixed."
-    );
-    assert!(
-        k3_curved_arrival < k3_floor_old,
-        "[#1026] the OLD full-ceiling floor {k3_floor_old:.4} should have demoted the curved K=3 \
-         arrival at EV {k3_curved_arrival:.4} — if it did not, this fixture no longer exercises \
-         the co-collapse bug and the regression is vacuous."
-    );
-    // Structure of the floor across K. K=1 keeps the original #1189 base gate
-    // (no co-collapse to forgive); for K >= 2 the per-atom share relaxes it BELOW
-    // the base (a curved K-atom fit is allowed to fall one atom's share short of
-    // the cumulative ceiling), and within the K >= 2 family the floor is
-    // NON-DECREASING in K (a larger dictionary is held closer to its ceiling),
-    // bounded above by the base #1189 gate it relaxes from.
-    let f1 = arrival_floor_k(k3_linear_ceiling, 1);
-    let f2 = arrival_floor_k(k3_linear_ceiling, 2);
-    let f3 = arrival_floor_k(k3_linear_ceiling, 3);
-    let f8 = arrival_floor_k(k3_linear_ceiling, 8);
-    assert!(
-        f2 <= f1 + 1e-12,
-        "[#1026] K=2 floor {f2:.4} should relax BELOW the K=1 base gate {f1:.4} \
-         (the per-atom share must forgive one collapsed atom)."
-    );
-    assert!(
-        f2 <= f3 && f3 <= f8,
-        "[#1026] per-atom-share floor is not monotone non-decreasing across K>=2 \
-         (K=2 {f2:.4}, K=3 {f3:.4}, K=8 {f8:.4})."
-    );
-    assert!(
-        f8 <= f1 + 1e-12,
-        "[#1026] the share floor exceeded the K=1 base gate at large K \
-         (K=8 {f8:.4} > base {f1:.4})."
-    );
-
-    // Guard the sentinel constant the fix exists to avoid pinning the loop at.
-    assert_eq!(SAE_FIT_DATA_COLLAPSE_COST, 1.0e12);
-}
